refactor: LottoNumber.kt 패키지 이동
