+++ src/main/kotlin/lotto/domain/LottoNumber.kt
@@ -1,9 +1,8 @@
-package lotto.component
+package lotto.domain
 
-const val START_LOTTO_NUMBER = 1
-const val LAST_LOTTO_NUMBER = 45
 
-private fun Int.isValidLottoNumber() = this in START_LOTTO_NUMBER..LAST_LOTTO_NUMBER
+
+private fun Int.isValidLottoNumber() = this in LottoNumber.START_LOTTO_NUMBER..LottoNumber.LAST_LOTTO_NUMBER
 
 @JvmInline
 value class LottoNumber private constructor(
@@ -11,6 +10,9 @@ value class LottoNumber private constructor(
 ) {
 
     companion object {
+        const val START_LOTTO_NUMBER = 1
+        const val LAST_LOTTO_NUMBER = 45
+
         fun from(number: Int): LottoNumber {
             if (!number.isValidLottoNumber()) {
                 throw IllegalArgumentException("로또 숫자는 1 부터 45 사이여야 합니다.")
